problem detection of services implemented.

+++ api/src/DownTrack.Application/Services/DepartmentServices.cs
@@ -20,13 +20,24 @@ public class DepartmentServices : IDepartmentServices
 
     public async Task<DepartmentDto> CreateAsync(DepartmentDto dto)
     {
+        var existingDepartment = _departmentRepository.GetById(dto.Id);
+        if (existingDepartment != null)
+        {
+            throw new ConflictException($"ID '{dto.Id}' is being used by another department.");
+        }
         var department = _mapper.Map<Department>(dto);
         await _departmentRepository.CreateAsync(department);
         return _mapper.Map<DepartmentDto>(department);
+
     }
 
     public async Task DeleteAsync(int dto)
     {
+        var department = _departmentRepository.GetById(dto);
+        if (department == null)
+        {
+            throw new ConflictException($"Department No.'{dto}' does not exist.");
+        }
         await _departmentRepository.DeleteByIdAsync(dto);
     }
 
@@ -35,13 +46,18 @@ public class DepartmentServices : IDepartmentServices
         var department = await _departmentRepository.ListAsync();
         return department.Select(_mapper.Map<DepartmentDto>);
     }
-    
+
     public async Task<UpdateDepartmentDto> UpdateAsync(UpdateDepartmentDto dto)
     {
         var department = _departmentRepository.GetById(dto.Id);
+        if (department == null)
+            {
+                throw new ConflictException($"Department '{dto.Id}' does not exist.");
+            }
         _mapper.Map(dto, department);
         await _departmentRepository.UpdateAsync(department);
         return _mapper.Map<UpdateDepartmentDto>(department);
+
     }
 }
